add wait loop for role definition

+++ scripts/delete-az-environment.sh
@@ -5,12 +5,12 @@ _BLUE='\033[1;34m'
 _RED='\033[1;31m'
 _NC='\033[0m'
 
-_FORMAT="$0 resourceGroupName"
+_FORMAT="$0 resourceGroupName [--deleterg]"
 
 _BASENAME=aibBuiUserId
 
 # arguments check
-if (( $# != 1 )); then
+if (( $# < 1 )); then
     echo -e "${_RED}Missing arguments.${_NC}\nUse this format:\n\t${_BLUE}$_FORMAT${_NC}"
     exit 1
 fi
@@ -33,5 +33,7 @@ az role definition list --query "[*].{scopes: assignableScopes, roleName: roleNa
 echo "Deleting user identities in $imageResourceGroup"
 az identity list -g $imageResourceGroup --query "[?starts_with(name,'$_BASENAME') ].id" -o tsv | xargs -I{} az identity delete -g $imageResourceGroup --ids {}
 
-echo "Deleting $imageResourceGroup"
-az group delete -n $imageResourceGroup -y
+if [ $2 == '--deleterg' ]; then
+    echo "Deleting $imageResourceGroup"
+    az group delete -n $imageResourceGroup -y
+fi
